backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
PyMuPDF
langchain
pinecone
sentence-transformers
groq
websockets
deepgram-sdk==3
Pillow
torch
transformers
sentencepiece